+++ hardware_interface/src/component_parser.cpp
@@ -177,7 +177,7 @@ std::size_t parse_size_attribute(const tinyxml2::XMLElement * elem)
   std::regex int_re("[1-9][0-9]*");
   if (std::regex_match(s, int_re))
   {
-    size = std::stoi(s);
+    size = std::stoull(s);
   }
   else
   {
